[clang] Implement CFG construction for @finally
Does for @finally what D112287 did for @try and @catch:
With this, clang can emit -Wreturn-type warnings for functions containing
@try/@finally, and -Wunreachable-code also works for those functions.

Conceptually, the @finally block happens after the statements in
the @try body, and after each @catch body.

Conceptually, the successors of the @finally are:

1. The block after the @try/@catch/@finally, for normal control flow
2. The end of the function, for exceptional control flow

Point 2 isn't 100% right: If a @try/@catch/@finally is nested inside
another @try, the successor of the @finally's exceptional edge
is that outer try's exceptional entry block (which then goes to
the outer @try's @catch blocks, which go to the outer @try's
@finally.)

But it's even more hairy than that. Consider:

    @try {
      return 1;
    } @finally {
      f();
    }

Every jump from inside a @try/@finally to outside the @try
needs to instead jump to the @finally first, which then gets
the jump target as a successor.

But it's even more hairy than that, because Objective-C allows control
transfer statements in @finally bodies. This is valid code:

    @try {
      return 1;
    } @finally {
      return 2;
    }

(The semantics are that the transfer of control in the @finally wins.)

(XXX: mention how this is different from the handling of dtors.)

(XXX: mention how checkers need to treat the CFG blocks of @finallys.)

Part of PR46693.