fix cfg construction for try with empty body with finally
The empty body caused a nullptr result from VisitObjCAtTryStmt(),
which led to the last non-null block in VisitCompoundStatement()
being the return.

For now, return the finally block if there's no try body block.
It might be better to always create a block for the try, even
if it's empty, similar to the

      // We can reach here if the "then" body has all NullStmts.
      // Create an empty block so we can distinguish between true and false
      // branches in path-sensitive analyses.

in VisitIfStmt() code.

XXX what if the finally has an empty body?

XXX do we have this bug in the other Visit*Try() functions
(or elsewhere)?